requirements.txt
streamlit
numpy
joblib
plotly.graph_objects
scikit-learn
